requirements.txt
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bcrypt==5.0.0
beanie==1.25.0
billiard==4.2.4
celery==5.3.6
certifi==2026.2.25
cffi==2.0.0
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==42.0.5
Deprecated==1.3.1
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.111.0
fastapi-cli==0.0.24
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.27.0
idna==3.11
Jinja2==3.1.6
kombu==5.6.2
lazy-model==0.2.0
limits==5.8.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
motor==3.3.2
orjson==3.11.7
packaging==26.0
passlib==1.7.4
prompt_toolkit==3.0.52
pyasn1==0.6.3
pycparser==3.0
pydantic==2.7.1
pydantic-settings==2.2.1
pydantic_core==2.18.2
Pygments==2.19.2
pymongo==4.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.3
redis==5.0.4
rich==14.3.3
rich-toolkit==0.19.7
rsa==4.9.1
sentry-sdk==1.45.0
shellingham==1.5.4
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
starlette==0.37.2
structlog==24.1.0
toml==0.10.2
typer==0.24.1
typing_extensions==4.15.0
tzdata==2025.3
ujson==5.12.0
urllib3==2.6.3
uvicorn==0.29.0
uvloop==0.22.1
vine==5.1.0
watchfiles==1.1.1
wcwidth==0.6.0
websockets==16.0
wrapt==2.1.2
